backend/requirements.txt
# Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Simulation Engine
numpy==1.26.3
scipy==1.12.0

# LLM Agent
langchain==0.1.4
langchain-google-genai==0.0.6
langsmith==0.0.83
langgraph>=0.0.20
langchain-core>=0.1.16

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Utils
python-multipart==0.0.6
httpx==0.26.0

# CORS
starlette==0.35.1
